moved auth to root
