+++ .env
@@ -1,6 +1,6 @@
 REACT_APP_ENTRY_POINT=los
-REACT_APP_API_URL=https://digix.aiplsolution.in/ratnaafin/
+# REACT_APP_API_URL=https://digix.aiplsolution.in/ratnaafin/
 # REACT_APP_API_URL=http://10.55.6.72:8081/
-# REACT_APP_API_URL=http://10.55.6.62:8081/
+REACT_APP_API_URL=http://10.55.6.63:8081/
 REACT_APP_DEBUG_MODE=false
 REACT_APP_MOCK_API=false
